core-framework/requirements.txt
# 核心依赖
pydantic>=2.0.0
loguru>=0.7.0
pyyaml>=6.0

# 异步支持
asyncio-mqtt>=0.13.0

# 可选依赖（用于特定功能）
redis>=4.5.0
psycopg2-binary>=2.9.0

# 开发依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pre-commit>=3.0.0